fix: ios: SignatureReady screen
